# Optimized SHA-256 Hash Function for Performance and Parallelism